[feat] map 파일 삭제 및 panTo 적용
